Match-id-53f14988f068811ac0a0114b824838d0311196ba

+++ app-engine/services/knowledge-base/src/main/java/com/huawei/jade/app/engine/knowledge/dto/KbVectorSearchDto.java
@@ -36,5 +36,5 @@ public class KbVectorSearchDto {
     /**
      * 阈值
      */
-    private Double threshold;
+    private Float threshold;
 }
